[REVIEW] Mapper: format console output, order mapping line to later, comment code, make check for empty string more obvious
The "re-mapping" line was ordered after the other checks because
otherwise two lines would be logged, leading to a confusin output.

The logging was extended to mention the names.

The first check was commented as it was not obvious to me after reading
it multiple times.

The check for an empty mapping target was changed to test for an empty
string because it would otherwise be potentially redundant to the
previous one. Checking for the naming pattern would be better maybe.